requirement.txt
aioservice
aiohttp_jinja2
python-picassoclient
python-functionsclient
